Images for project
Please enter all images here so that we can view then and they can be used going forward.
Perhaps we should start character images with "char" and background images with "bg" followed by the title of the image. 
Images when replaced should be named the same, version of the older images should be kept in the google drive for the team.